requirements.txt
discord.py
python-dotenv
langchain
langchain-google-genai
google-generativeai
Pillow
aiohttp
requests
